ToxicBot/requirements-dev.txt
discord.py==1.5.0
tensorflow==2.3.1
Keras==2.4.3
psycopg2-binary==2.8.6
pre-commit==2.7.1
flake8==3.8.4
black==20.8b1